Make trap defense rules consistent

+++ mods/cards/BN6_Standard_191/entry.lua
@@ -157,7 +157,9 @@ function card_init(user)
 
 		defense_rule.on_replace_func = uninstall_all
 
-		defense_rule.defense_func = function(_, _, _, hit_props)
+		defense_rule.defense_func = function(defense, _, _, hit_props)
+			if defense:damage_blocked() then return end
+
 			if hit_props.element == Element.Cursor or hit_props.secondary_element == Element.Cursor then
 				uninstall_all()
 			end
